Adjust for Flowdock adapter connected event oddity
The adapter currently emits the connected event a bit early, so it's not
quite time to send a message yet. We have a PR out on the adapter to fix
this, but in the meantime we special-case robot.adapter and do some nutty
stuff to catch the connection when it's ready.