Add initial Zeplin 'API' wrapper
The API supports a zeplin.Session, defined with a username and password,
and interacts both with Zeplin's unpublished API and in some cases with
API data emitted directly into HTML pages to expose information about
projects, screens, and comments (both 'dots' and their replies).

API data has associated types so we know what we're expecting to be
working with.