Add Zeplin->Flowdock integration
The integration creates threads for new screens, as well as posting
comments into their respective screen threads as they come in. A preview
of the screen is embedded in the thread, as well as metadata about it.
Comments and their resolutions all link back to the appropriate entry on
the screen's page, and the thread's link is to the screen.