Added basic Zoom interaction API
Right now, the API supports getting a Zoom session (that is
known-working, by doing a request for all available users in the Zoom
account), which then allows requesting the next available meeting.
This is done by checking all available users for all users with no
current live meeting, and then randomly choosing an available user to
create a meeting for.