Fix bug in getting flow id from name
Without the check for robot adapter flows, scripts fail at load time
Can not check length of undefined

Also added logging here in the case that room name is returned
instead of the id. Returning the name works in some cases, like
developing in the shell, but not in other cases, like trying to
send a message to another room. Envelope expects flow id

There is probably a better solution here, but logging is at least
a way to point back to this issue if it comes up again