Remove schedule command from access limit list
Its presence in this list has been a lie all along, since the command
does not have a named id assigned to it.
The schedule command has already lived past the trial phase in which
we would want to restrict it to these rooms anyway, so removing it
from the list instead of fixing it by adding an id.