Start Heimdall without Flowdock
Here we removed the Flowdock adapter from the Heimdall Docker entrypoint
to do a couple things.  First we wanted to startup Heimdall in
thesis-ops without enabling Flowdock, just to see if it would boot.
Second, we used it as an example to talk about modifying run commands
for Docker containers.