Use image in config file and set command
To test deployment without going through the Circle build pipeline we
baked an image on local and pushed it manually to the thesis-ops
container registry.  We fetched this image via the image parameter in
config.  Additionally, we went through an exercise of setting command to
show how we can override the Dockerfile entrypoint without having to
bake a fresh image every time we want to change the run command.