First pass, basic keep-test faucet
We run a handful of internal ethereum chains, they're a pain in the ass
to fund accounts on.  You have to use the geth console, or geth CLI to
transfer eth.  Our keep-test environment has a public transaction node.
This means we can take advantage of Heimdall to fund accounts on this
chain, instead of having to use the aforementioned methods.  Here we
provide a very basic Heimdall command that funds a passed account with
10 eth.