Rename faucet script
Rather than peg the script to an environment we're making the naming
generic, with a note that the script only works for keep-test at the
moment.  This is to set the intention that we should be able to use this
script for multiple environments, it just needs to be extended.