Remove funded log
tl;dr here is that we think the ordering issues on Flowdock are due to
the fact that the unlocked and funding msg.sends are firing at about
the same time, and that Flowdock doesn't guarantee ordering in that
case.  We're just going to go ahead and remove the unlocked log, as this
falls over if the accounnt isn't unlocked anyway (and we have logging
for that).