Add withConfigOrReportIssues via TypeScript conversion
The TypeScript conversion brings a couple of additional changes.
fetchConfigOrReportIssue and withConfigOrReportIssue stop relying
directly on a robot, instead relying on an `issueReporter`, which takes
an error message and reports the issue as the issue reporter wishes.

withConfigOrReportIssue introduces a way to group many
fetchConfigOrReportIssue calls and require that they all succeeds before
running some additional code. github-auth and zoom are both ported to
use it instead of directly calling fetchConfigOrReportIssue. They no
longer set up their respective Hubot responders if the required
configuration isn't available.

Last but not least, we gain an issueReporterForRobot that wraps a robot
into a function suitable for passing as an issue reporter for
fetchConfigOrReportIssue and withConfigOrReportIssues.