Fix bugs in updated calls to processTemplateString
In the lastUrl function, fix the attempt to get the lastUrl from the
currently running job. That property is stored in the job metadata,
not directly on the job object

When initially calling processTemplateString we do not yet have a job
created. We call it at this point to force the command to error in
a way that will be visible to the user if the template string usage
is incorrect
Otherwise it would fail only when the job is running, and show up
only in our logs, and be invisible to the user

In order to still call it at this stage of the job creation, we have
to pass an empty object

Because the empty object does not have metadata, we need to update
the lastUrl function to explicitly check for metadata before trying
to pull a lastUrl property from it