Reduce or eliminate rate-limiting in Zoom meeting lookup
The way Zoom meeting lookup was written, each account fired off a
roughly simultaneous set of requests to Zoom to check its availability,
irrespective of whether or not it would be used. Additionally, the use
of Promise.all meant that if *any* of those requests failed, the whole
function failed.

On the API end, Zoom would quickly rate limit once a certain number of
accounts were being checked, as it would be hit by 30+ requests in a
short period of time.

On the UX side, because all accounts were checked and *then* filtering
was performed, starting a new zoom meeting required waiting for up to 3
requests to complete per account, which can get slow, particularly when
rate limiting or control kicks in.

The new approach sorts pro vs basic and randomizes each list. This gives
close enough to the same random choice effect of the previous mechanism,
plus starts with only pro accounts. The rest is done in a regular for
loop with awaits, so that each account is checked in turn instead of all
accounts always being checked.

The result should be faster on average, and should only trigger rate
limits in very outside cases.