Move deploy steps to a callable workflow
This allows the main build not to reference the deployment environment,
so that the latest deployment is reflected correctly in GitHub. It also
prepares the deploy workflow to be better structured for reuse by other
projects.

Additionally, the Dockerfile no longer runs tests (which is redundant to
the test workflow), and the image push still happens as part of
building. Builds also depend on successful tests.