Update static IP for web ingress
The new IP is owned by the correct (thesis-ops) project, and is being
hooked up to a proper subdomain in Cloudflare. The kube secrets map has
been updated to point to that subdomain so that once this config is
applied the valkyrie pods can be kicked and pick up the updated hostname
info.