Add new reminder listing, updating and canceling
Updates can be to recurrence or message, and the verbs can be update or
edit, and for deleting can be delete/remove/cancel. Additionally,
`remind` can be used as a shorthand for `reminder` to avoid confusion
between ?remind for creating and ?reminder (list/update/etc) for acting
on the reminders.

All jobs get an id that increments over time, and the scheduler tracks
the max id it has seen. The concept of a PersistedJob is introduced that
includes the id, vs a regular Job that may not. Jobs are tracked by id
in addition to in the master list, and canceling or updating a job has
the opportunity to restart the schedule loop to accommodate the
possibility of the "next" job changing.

Additionally, updating a job results in removing and re-adding it so
that it is kept in the same order. This is potentially slower than
adjusting its position in-place, but it's simpler to understand and at
the job amounts we're expecting it shouldn't be a huge issue.