Fix next occurrence on day of recurrence
If a weekly next occurrence was computed on the day of the recurrence,
it would always schedule out by the weekly interval. This meant that
asking for every Monday at 4pm on Monday at noon would not trigger a
reminder until the following Monday.

The same was true for monthly next occurrence on the same day of the
month, and is fixed as well, except this one had the specs aligned
incorrectly as well—they are now correct.