Update _alias-help-command to load properly
This was renamed to .cjs during the TS port, but ultimately the approach
that used the .cjs extension was not used, and .cjs extensions weren't
being loaded.

The _alias-help-command allows e.g. `\remind help` to work as if you
typed `\help remind`, tapping into the existing help script that
extracts help info from script headers.