Add timezone command for users to set a timezone preference
The timezone command tells the user the timezone they currently use, and
they can pass a timezone to it to use a specific timezone for reminders
and potential future timezone tidbits.

This is all in a scripts/user-preferences script that will be the home
for any user-specific preferences the system will allow in the future.

Because certain timezones are weird (e.g. with daylight savings time),
the script allows for aliases that resolve to a more specific technical
timezone.