Add timezone support to remind command
The remind command now resolves the user's current timezone (if they
have set one) and uses it to manage the times they specify. These times
are then adjusted to UTC for storage, and adjusted back for display to
the user.

Note that the jobs themselves don't store the timezone that was used
when the job was scheduled, times are always presented in the timezone
of the requesting user.

Note also that dayjs has certain bugs when doing math with timezoned
objects. These seem to mostly be related to a non-UTC system timezone,
but regardless a move to luxon will be necessary to remove this class of
bug possibilities.