Move dayjs to luxon
Luxon uses the underlying Intl API for handling timezones, etc, and is
therefore less buggy than dayjs's timezone handling.

Tests are moved to always use UTC as there were baked-in assumptions in
the expectations that things would be interpreted as UTC. Explicit
timezone tests should also be added.

Dependency on chrono-node is also dropped as it is no longer used, and
was pulling in another version of dayjs.