Load existing reminders in the scheduler
Previously, existing reminders were being read in the remind script,
using the script's reminder key (`hubot_reminders`). The job scheduler,
meanwhile, was persisting to the default key it was using, which was
`jobs`.

Because the job scheduler is already in charge of persisting, and
because the job scheduler is quite intertwined with the robot already,
this commit moves all responsibility for robot brain interaction into
the scheduler. This means the scheduler itself now reads the existing
job list from the robot brain at instantiation time, in addition to the
existing responsibility of writing updated jobs to the brain when they
change.

Finally, moving this responsibility into the scheduler means the
scheduler now uses the same key to read and write jobs in the robot
brain.