Fix a missing useFakeTimers in reminder specs
This also required re-fixing an off-by-one error in expected test
results. Why didn't the off-by-one error get caught? Luxon at first
glance looks like it deals with numbering around weekdays differently
than dayjs, but in fact it deals with them the same! Thinking that it
treated them as 1-based instead of 0-based (somewhat but not really
true), the off by one error in the test result felt like it was due to
the change in library. In fact, it was because we were using the local
date, which happened to be a Saturday, against a 6-day offset. The
original test was on a Friday against a 6-day offset. In one case the
resulting weekday was 11, in the other 12.