Update Dockerfile to include node-canvas libraries
node-canvas is used to render room avatars. It requires certain native
dependencies like cairo and pango, so the dockerfile needs to set these
up both for the `yarn install`, which will use them to build the
`node-canvas` bridge, as well as for when the container is actually
running.

For simplicity, the Dockerfile installs the relevant libraries using
`apk` for the runtime stage, while installing the dev versions for the
build stage. The other option would be to only copy over the appropriate
so/lib from the build stage to the final stage, but this would require
checking all dependencies to copy just the relevant files. The package
manager seems easier, even though the resulting image might be a bit
bigger (~3x compared to before, unfortunately).