Generate and set room avatars when taking admin
Avatars are set to the room alias followed by `.png` as a filename, and
uploaded to the matrix file store. They are drawn using node-canvas,
which unfortunately has native dependencies but will do the trick
overall---the non-native-dependency versions of the canvas API didn't
seem to support the right text rendering.

The design for avatars is taken from the Thesis Coda/Element design
system.