Fix alias setting for new aliases
Matrix requires creating an alias as a separate thing from sending the
canonical alias event. The canonical alias event just announces the
room's alias to its members...

Allow for errors in the alias setting, log them, and don't mention the
alias in the admin takeover message if the alias failed to be set.