Port everything to ESM and Thesis hubot fork
This fork moves TypeScript and CoffeeScript support directly into Hubot
itself, as well as moving the whole system to ES Modules. Some
experimental flags are required for this, but the hubot package takes
care of that, while valkyrie can focus on leveraging the advantages.

The changes are mostly around using `export default` instead of
`module.exports`, using file extensions in non-package-name imports, and
importing certain modules as defaults instead of using `* as`.