Fix Jest execution for ES Module Valkyrie
Jest execution required some light massaging of the Jest config (which
is now in TypeScript for completeness), but also required bumping the
TypeScript version to 5.0.0-beta, as this was a prerequisite for
enabling the `allowImportingTsExtensions` tsc flag. This was required
to allow compiling and type-checking the TypeScript files (which the
jest config does), since otherwise the TypeScript compiler will complain
that you cannot add the `.ts` extension to imports. We do, because the
new TypeScript config in hubot understands and transpiles these files
directly, rather than requiring us to compile them ourselves.

Score one for the good guys, eventually.