Fix trailing linting issues
Now that the eslintrc is loading, a few lint issues came up that can
be fixed.

First, the import/extensions config is a mismatch for the ES Module
loader strategy we're using, where explicit file extensions are
necessary for non-package imports. Additionally, this surfaced one
import that still wasn't using the extension explicitly.

Lastly, an error around the use of a RegExp constructor instead of a
literal is now tamed.