Fix lifecycle for Discord script attachment
There is a possibility of a race condition with hubot boot, where the
script could be loaded before the adapter is available, but if it is
loaded after the adapter is available the `connected` event may already
have fired.

We now hook into the `adapter-initialized` event if and only if the
adapter is undefined when the script runs, and attach scripts at that
time. If it is already defined, we immediately attach the scripts.