Add first-pass Figma integration
The Figma integration provides a /figma command that allows connecting
and disconnecting named teams to a given channel. Figma webhooks are
created for these teams, notifying of comments, file updates, and file
version saves (for named versions).

The code is quite ugly at the moment, and is all in one big blob of
Discord script.