Improve Figma file update notifications
File updates are now tracked over time. We wait until a file has not
seen updates for 10 minutes, at which point the file update is posted
based on the latest version in the version history for the file. That
version is also used for the embed's author information.

If the latest version in the version history is named version no update
is posted as it's expected the FILE_VERSION_UPDATE event will handle it.