Add comment triggers for tests and deploys
The overall goal is to allow temporary deployments of non-main builds to
the production valkyrie for final testing, or for debugging particularly
thorny behaviors. The way to trigger this is to post an explicit request
for the deployment in a PR comment including the text `deploy, please`.

A new job, decide-parameters, sets parameters based on the current build
environment to determine when builds, tests, and deploys should run.
This job is used to trigger the proper behaviors on issue comments.

This also simplifies the logic for if statements for deploys, which were
having to check refs against multiple conditions in the build job
bodies; those checks are now replaced with one check against the
parameter for deploying.

This also allows builds to run independent of tests in cases where a
deploy is requested via comment.