Add initial implementation of thread archive checks
These nudges happen when a thread is within 24 hours of its auto-archive
time frame, and nudge the participants to take action to close the
thread out.

The current iteration gives the option to immediately archive, ask for a
decision, ask for a status, or ask for a task to be captured. The three
latter actions prompt for a person to identify as the person to decide,
give status, or capture task. The robot then pings that person and lets
them know that the thread will be auto-archived in 24 hours and to give
their info in that time frame.

Notably, by this point the bot has already set the thread to auto-
archive in 24 hours.

Soon we'll want a couple of these paths to re-extend the auto-archive
time frame back to the server default if possible, once the designated
person replies.

There is also a currently-unused code path that heuristically determines
if the full lifetime of a thread has been synchronous; in that case, it
does not act in any way to change auto-archive or prompt for outcomes.
These threads are treated as representing complete conversations.