Drop confusing everyone tagging in thread-management
Originally thread-management auto-tagged everyone on general/main
channels, but this doesn't actually have the desired effect in threads
(@everyone in a thread only tags the people already joined to the
thread).

Drop that behavior since it's confusing to have something that looks
like it has an effect that it doesn't actually have.