Add twitter embed fixer
When a new twitter.com/x.com link is seen, Valkyrie auto-replies with
the same link switched to fxtwitter.com (see
https://github.com/FixTweet/FxTwitter for the implementation). This
produces an actual Discord embed, as well as auto-expanding t.co links
and doing a couple of other useful things.

Those who wish to can also change the link to always redirect to nitter
instead of twitter (e.g. if you do not stay logged in or if twitter is
broken for you).