Don't insert a space for privacy-tagged non-thread channels
For thread channels, the space after the lock emoji makes sense as the
titles are usually sentences, and require some visual space. For
non-thread channels, the space after the lock emoji becomes a - instead
as these channels don't support spaces.

To fix this discrepancy, heuristically check for thread names with
spaces or capital letters and don't insert a space after the emoji for
these.