Move channel metadata handling to a common file, add permitted features
Permitted features are tracked for a thread, channel, or category. The
functionality for whether a thread is sync or not is rolled into a
generic channel metadata functionality, which now includes permitted
features per thread/channel/category.

Note that there is not yet an external way to add to the permitted
features metadata.