Channel Permission set
This commit adds a new `channel-management` helper to Valkyrie which will always ensure permissions on new channels within the `defense` category are assigned to the same permission set as the channel category permissions.

It then sends a message to the newly created channel, listing the permissions and access levels they have.