Add ext-*-audit channel invite
This adds a implementation into the invite script which generates an invite to the specific channel based of a channel being made in the defense category with `ext-*-audit` format.

To-do: Link with role-manager to whirl up a role with the same channel name, granting access to that channel only.