Add /create-invite command
This commit adds a new discord command `/create-invite` that allows you to generate a new invite within the channel that the command is run from.

Max age / max uses: Included are two parameters, `max-age` and `max-uses` that optionally let you specify the amount of time the invite link will be available for, and the number of invites. By default this will be set to 24 hours and for only 10 uses.