Create ext-*-audit channel role
Once a new `ext-*-audit` channel has been created, this will now create a role with the exact same name as the channel, and assign permissions to only view that channel.

Now will need to get a role assignment going when a new user joins the server based off the invite.