Custom auto-tagging
### Notes
This adds the ability to modify the auto-tagging function of Valkyrie based off of a defined list of `PRIVATE_CHANNELS`. It also will match the tagging to the role specified, in order to have custom roles tagged for specific threads started in a channel.

### Handling IDs
Alternatively, we can also modify this to work directly with channel / role IDs rather than names, but in testing it seems to be working as it should.