Channel Archiver
### Notes
This WIP PR focuses on building a channel archiver function into Valkyrie. So that you can easily dump channel logs from a specific channel at point of archive.

Due to Discord APIs message rate limit of 100 messages, we run a loop depending on the total number of messages in the channel, then combine into a file which is sent back to the channel.

### WIP
This is just the initial test setup of this, some things we need to add still are:

- Better storage for the archival message file (dotenv is temp)
- Message headers specifying date of archive, channel name, etc
- Permission flags, access to archive
- Introduction of a command, valkyrie interface, etc