Update auto-join to try channel name prefixes
In particular, this allows categorizing not just by top-level category
in Discord, but also refining by a channel prefix. As an example,
a channel #mezo-engineering-musd in the Mezo category would previously
only auto-join the roles "Mezo Engineering mUSD" and "Mezo", because one
would match the channel and one the category.

The new approach tries all prefixes of the channel name, so that "Mezo
Engineering" can also be auto-tagged.

With the current Discord channel name structure, this means most items
will be auto-tagged in this phase instead of by category, since project
channels always prefix their project name. One key exception is Studio.