Implement embeds by issueId
### Notes
This implements a new feature to embed linear links based off of the `issueId` if it's placed in a message, such as `THESIS-123`.

Additionally this adds a function to delete the embed if the original message link is deleted / changed.

Still need to implement editing the original embed if it changes.