Fix role lookup for autotagging
The previous version was searching for the first role that matched
either prefix match, so if Mezo happened before Mezo Commerce (for
example) in the role list, it would match.

This version checks for matches for all of the prefixes, and then pulls
the most specific prefix that matched (by virtue of array ordering,
since prefixes are stored most- to least-specific).