Fix bug with linear updates disconnect
### Notes
This fixes is a bug that was happening where the /linear-updates disconnect command was not working properly due to the incorrect field mapping. Now the proper flow returns the connected team and also gives a message if the team is not connect to the channel when trying to disconnect it.