Resolve Linear updates issues with Edited & length
### Notes
This resolves an issue where a Linear project update is reposted when an update is made to an existing update. If the request has updatedFrom, we can exclude this from being posted.

In addition, this caps the description text to 500 characters and excludes any text in the embed that is under a collapsable section with +++